Partial commit for supporting various types of analytics card.
We will come back to this. But now we need to resolve the data 
dependency clutter.

Some cards will rely on last four weeks data. We need to express these 
data dependencies that do not block rendering of our main metric – time 
left for work today. Because it does not depend on the aggregated data.