Remove the mocked values for
-  now
-  today
-  start of week

Experimented with mock date, which changes the native date object to 
whatever you like.

So, after this,
`MockDate.set("2019-04-03T09:00")`

the `moment()` will return a timestamp representative of the argument 
passed to `MockDate.set`

And it worked well.

Will know try to create a story with a knob which can mock *now*.