Mechanism to change datetime from storybook
This will provide us a way to quickly check our time left for work 
across different date times and events configuration

For events, we are mocking the `FirestoreDataContext`. One configuration 
provides no event and one configuration provides events for the first 
week of April

So, if you want to test the `TimeLeftForWork` component during the first 
week of April

-  select the appropriate context from the storybook toolbar
-  change the date time in knobs to any time in first week of April like 
2019-04-02T10:00

Important Note: The context addon is in alpha. Changing the knob value 
does not automatically update the story. Select, Note and then Canvas 
again to update the story